BUGFIX * vstudio7 project file for AAI * fixed AAI build for vstudio7 * updated installer * changed version to 0.73b1 and project name on load screen to "Spring" instead of "TA Spring"
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2061 37977431-3df6-0310-b722-df95706aa16b

+++ AI/Global/AAI/AAIBrain.cpp
@@ -474,7 +474,7 @@ bool AAIBrain::ExpandBase(SectorType sectorType)
 			dist = 0.1;
 			for(list<AAISector*>::iterator sector = sectors[0].begin(); sector != sectors[0].end(); sector++)
 			{
-				dist += 2.0 * sqrt(pow( float((*t)->x - (*sector)->x) , 2.0) + pow( float((*t)->y - (*sector)->y) , 2.0));
+				dist += 2.0 * sqrtf(powf( float((*t)->x - (*sector)->x) , 2.0f) + powf( float((*t)->y - (*sector)->y) , 2.0f));
 			}
 
 			if(sectorType = LAND_SECTOR)
